Reading main CRM block completed
Porting CRM reading code from native assembly to editor assembly for
better maintainability. This commit includes the "ReadMainBlock" method
which is where the bulk of the reading takes place. Needs review to
ensure nothing is broken, especially the functions reading image files.
Everything is in "Editor\AGS.Editor\RoomLoader.cs", pending review and
distribution to appropriate files and/or classes.